+++ tests/test/perl/range-re.pl
@@ -1,0 +1,11 @@
+# range/flip-flop: second regexp after .. must be recognised
+/'/ .. /"/
+print if /'/ .. /"/;
+print if /foo/../bar/;
+print if /start/ ... /end/;
+
+# still a regexp after other operators
+print if /a/ or /b/;
+
+# still division, not a regexp
+$a = $b / $c / $d;